Fix type mismatch when handling uploaded file data
Update `lib/object-storage.ts` to cast `result.value` to `Uint8Array` before passing it to `Buffer.from()`, resolving a TypeScript type error.

+++ lib/object-storage.ts
@@ -22,7 +22,7 @@ export const objectStorage = {
             if (!result.ok) {
                 return null;
             }
-            return Buffer.from(result.value);
+            return Buffer.from(result.value as Uint8Array);
         } catch {
             return null;
         }
